feat(tenancy-p0): KekVersion smallint NOT NULL DEFAULT 1 + short call sites
- TammaModelConfiguration: Property<int?> → Property<short>.HasDefaultValue((short)1)
- AdminTenantsEndpoints (×2), LruPooledTenantConnectionResolver: EF.Property<int?> → (int?)EF.Property<short>
- KekCabinetHealthCheck (×2): same EF.Property conversion (null branch becomes dead code, CS0472 warning expected)
- KekRotationCoordinator (×3): query predicate drops ?? 0 coalesce; write casts to (short)toVersion
- EncryptAndPersistConnectionStringActivity: read via (int?)(short?) chain; write casts (short)kek
- SoftDeleteTenantRowActivity, EmitDeletedSuccessActivity: drop null-set of NOT NULL column (dead on delete path)
- All test seeder sites: CurrentValue = (short)1 / (short)kekVersion
- KekCabinetHealthCheckTests: seeder uses NpgsqlParameter<short>, legacy-NULL tests marked [Ignore] with rationale
- Migration: TenancyP0_KekVersionSmallint — AlterColumn smallint NOT NULL defaultValue=(short)1

+++ apps/tamma-elsa/tests/Tamma.Api.Tests/Epic28/LruPooledTenantConnectionResolverTests.cs
@@ -83,7 +83,7 @@ public class LruPooledTenantConnectionResolverTests
         entry.Property("Status").CurrentValue = status;
         entry.Property("EncryptedConnectionString").CurrentValue =
             Encoding.UTF8.GetBytes(connectionString);
-        entry.Property("KekVersion").CurrentValue = 1;
+        entry.Property("KekVersion").CurrentValue = (short)1;
         await ctx.SaveChangesAsync();
         return tenantId;
     }
@@ -164,7 +164,7 @@ public class LruPooledTenantConnectionResolverTests
             var entry = ctx.Tenants.Add(tenant);
             entry.Property("Status").CurrentValue = "active";
             entry.Property("EncryptedConnectionString").CurrentValue = Array.Empty<byte>();
-            entry.Property("KekVersion").CurrentValue = 1;
+            entry.Property("KekVersion").CurrentValue = (short)1;
             await ctx.SaveChangesAsync();
         }
